move notification component from gallery to app
